Renamed py3_helpers to compatibility.
Missed py3_helpers renaming.
