+++ pyocd/utility/compatibility.py
